// frontend/src/types/ProviderProps.ts
export default interface ProviderProps {
    id: string;
    name: string;
    logo?: string;
    city: string;
    address: string;
    email?: string;
    phone: string;
    whatsapp?: string;
    facebook?: string;
    messenger?: string;
    observation?: string;
    createdAt: string;
    deletedAt?: string;
}